deleted old go codebase
