+++ go/decode.go
@@ -1,148 +1,115 @@
-package bufti
+package bufti2
 
 import (
 	"bytes"
 	"encoding/binary"
 	"fmt"
-	"strings"
+	"reflect"
 )
 
-// Decode decodes the specified byte array into a map.
-func (m *Model) Decode(b []byte) (map[string]any, error) {
-	if b == nil {
-		return nil, ErrNilSlice
+func (m *Model) Decode(data []byte, dest any) error {
+	v := reflect.ValueOf(dest)
+	t := reflect.TypeOf(dest)
+
+	buf := bufferPool.Get().(*bytes.Buffer)
+	defer bufferPool.Put(buf)
+
+	buf.Reset()
+	if _, err := buf.Write(data); err != nil {
+		return err
 	}
 
-	buf := bytes.NewBuffer(b)
-
-	version, err := buf.ReadByte()
-	if err != nil {
-		return nil, err
+	var version uint32
+	if err := binary.Read(buf, binary.LittleEndian, &version); err != nil {
+		return fmt.Errorf("%w: failed to read protocol version", ErrBuffer)
 	}
-	if int(version) != MajorVersion {
-		return nil, fmt.Errorf("%w: message uses version %d, this package uses version %d", ErrVersion, version, MajorVersion)
+	if version != ProtocolVersion {
+		return fmt.Errorf("%w: incompatible bufti version: this package uses version %d, buffer uses version %d", ErrVersion, ProtocolVersion, version)
 	}
 
-	return m.decode(buf, len(b))
+	return m.decode(buf, t, v)
 }
 
-func (m *Model) decode(buf *bytes.Buffer, limit int) (map[string]any, error) {
-	bufti := make(map[string]any)
+func (m *Model) decode(buf *bytes.Buffer, t reflect.Type, v reflect.Value) error {
+	var fieldCount uint32
+	if err := binary.Read(buf, binary.LittleEndian, &fieldCount); err != nil {
+		return err
+	}
 
-	for range limit {
-		var index byte
-		if err := binary.Read(buf, binary.BigEndian, &index); err != nil {
+	if t.Kind() != reflect.Pointer || v.IsNil() {
+		return fmt.Errorf("%w: dest has to be a pointer, instead: %s", ErrInput, t.Kind())
+	}
+	v = indirectValue(v)
+	t = indirectType(t)
+
+	switch t.Kind() {
+	case reflect.Struct:
+		return m.decodeStruct(buf, t, v, int(fieldCount))
+	case reflect.Map:
+		return m.decodeMap(buf, t, v, int(fieldCount))
+	default:
+		return fmt.Errorf("%w: invalid destination type %s", ErrInput, t.Kind())
+	}
+}
+
+func (m *Model) decodeStruct(buf *bytes.Buffer, t reflect.Type, v reflect.Value, fieldCount int) error {
+	fieldMap := make(map[string]reflect.Value)
+
+	for i := range t.NumField() {
+		field := t.Field(i)
+		value := v.Field(i)
+
+		if !value.CanInterface() {
+			continue
+		}
+
+		fieldName := field.Name
+		if tag := field.Tag.Get("bufti"); tag != "" {
+			fieldName = tag
+		}
+
+		fieldMap[fieldName] = value
+	}
+
+	for range fieldCount {
+		index, err := buf.ReadByte()
+		if err != nil {
+			return err
+		}
+
+		schemaField, exists := m.schema[index]
+		if !exists {
+			return fmt.Errorf("%w: index %d does not exist on model %s", ErrModel, index, m.name)
+		}
+
+		if err = schemaField.fieldType.Decode(buf, fieldMap[schemaField.label]); err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+func (m *Model) decodeMap(buf *bytes.Buffer, t reflect.Type, v reflect.Value, fieldCount int) error {
+	if t.Key().Kind() != reflect.String || t.Elem().Kind() != reflect.Interface {
+		return fmt.Errorf("%w: destination has to be a map[string]any, instead: %T", ErrInput, t.String())
+	}
+
+	for range fieldCount {
+		index, err := buf.ReadByte()
+		if err != nil {
 			break
 		}
 
 		schemaField, exists := m.schema[index]
 		if !exists {
-			return nil, fmt.Errorf("%w: index not found (%d)", ErrBufferFormat, index)
-		}
-		valType := schemaField.fieldType
-		label := schemaField.label
-
-		value, err := decodeValue(buf, valType)
-		if err != nil {
-			return nil, err
+			return fmt.Errorf("%w: index %d does not exist on model %s", ErrModel, index, m.name)
 		}
 
-		bufti[label] = value
+		var mapValue any
+		if err = schemaField.fieldType.Decode(buf, reflect.ValueOf(&mapValue).Elem()); err != nil {
+			return err
+		}
+		v.SetMapIndex(reflect.ValueOf(schemaField.label), reflect.ValueOf(mapValue))
 	}
-
-	return bufti, nil
-}
-
-func decodeValue(buf *bytes.Buffer, valType BuftiType) (any, error) {
-	var size int
-	if valType == "string" || strings.HasPrefix(string(valType), "list:") || strings.HasPrefix(string(valType), "map:") || strings.HasPrefix(string(valType), "model:") {
-		var v uint16
-		if err := binary.Read(buf, binary.BigEndian, &v); err != nil {
-			return nil, err
-		}
-		size = int(v)
-	}
-
-	switch valType {
-	case Int8Type:
-		var v int8
-		err := binary.Read(buf, binary.BigEndian, &v)
-		return v, err
-	case Int16Type:
-		var v int16
-		err := binary.Read(buf, binary.BigEndian, &v)
-		return v, err
-	case Int32Type:
-		var v int32
-		err := binary.Read(buf, binary.BigEndian, &v)
-		return v, err
-	case Int64Type:
-		var v int64
-		err := binary.Read(buf, binary.BigEndian, &v)
-		return v, err
-	case Float32Type:
-		var v float32
-		err := binary.Read(buf, binary.BigEndian, &v)
-		return v, err
-	case Float64Type:
-		var v float64
-		err := binary.Read(buf, binary.BigEndian, &v)
-		return v, err
-	case BoolType:
-		var v bool
-		err := binary.Read(buf, binary.BigEndian, &v)
-		return v, err
-	case StringType:
-		p := make([]byte, size)
-		if _, err := buf.Read(p); err != nil {
-			return nil, err
-		}
-		return string(p), nil
-	default:
-		listType, isList := isListType(valType)
-		if isList {
-			var list []any
-			for range size {
-				item, err := decodeValue(buf, listType)
-				if err != nil {
-					return nil, err
-				}
-				list = append(list, item)
-			}
-			return list, nil
-		}
-
-		keyType, valueType, isMap := isMapType(valType)
-		if isMap {
-			m := make(map[any]any)
-			for range size {
-				key, err := decodeValue(buf, keyType)
-				if err != nil {
-					return nil, err
-				}
-				value, err := decodeValue(buf, valueType)
-				if err != nil {
-					return nil, err
-				}
-				m[key] = value
-			}
-			return m, nil
-		}
-
-		modelName, isModel := isModelType(valType)
-		if isModel {
-			model, exists := registeredModels.get(modelName)
-			if !exists {
-				return nil, fmt.Errorf("%w: can not find the model %s", ErrModel, modelName)
-			}
-
-			pl, err := model.decode(buf, size)
-			if err != nil {
-				return nil, err
-			}
-			return pl, nil
-		}
-
-		return nil, fmt.Errorf("%w: invalid type (%s)", ErrBufferFormat, valType)
-	}
+	return nil
 }
