+++ go/go.mod
@@ -1,3 +1,3 @@
-module github.com/QYUbit/Bufti/go
+module bufti2
 
 go 1.22.5
