+++ go/literals.go
