--- go2/go.mod
@@ -1,3 +1,0 @@
-module bufti2
-
-go 1.22.5
